REFACTOR-#0000: move progress bar in file with base partition manager
